+++ projectify/templatetags/projectify.py
@@ -14,9 +14,11 @@ from django.utils.html import format_html
 from django.utils.safestring import SafeText, mark_safe
 from django.utils.translation import gettext_lazy as _
 
+from projectify.lib.types import SupportsGetAbsoluteUrl
 from projectify.lib.utils import static_image_get_with_dimensions
 from projectify.user.models import User
-from projectify.workspace.models import TeamMember
+from projectify.workspace.forms import WorkspaceSearchForm
+from projectify.workspace.models import TeamMember, Workspace
 
 logger = logging.getLogger(__name__)
 
@@ -33,7 +35,8 @@ def percent(value: Optional[float]) -> Optional[str]:
 
 @register.simple_tag
 def anchor(
-    href: str,
+    # Rename to path
+    href: Union[SupportsGetAbsoluteUrl, str],
     label: str,
     aria_label: Optional[str] = None,
     title: Optional[str] = None,
@@ -50,12 +53,16 @@ def anchor(
     """
     extra: Union[SafeText, str]
     a_extra: Union[SafeText, str]
-    if href == "":
-        raise ValueError("Empty href supplied")
-    try:
-        url = reverse(href, args=args, kwargs=kwargs)
-    except NoReverseMatch:
-        url = href
+    match href:
+        case "":
+            raise ValueError("Empty href supplied")
+        case str():
+            try:
+                url = reverse(href, args=args, kwargs=kwargs)
+            except NoReverseMatch:
+                url = href
+        case model:
+            url = model.get_absolute_url()
     # TODO, if we have a reverse match, we don't have external URLs
     # We could switch all callers of the anchor function to use the route name
     # and implicitly switch on external for all other URLs. After all, if it's
@@ -114,7 +121,7 @@ def circle_anchor(
     label: str,
     icon_style: str,
     # TODO rename to path, since we don't accept full URLs anymore
-    href: str,
+    href: Union[SupportsGetAbsoluteUrl, str],
     fragment: Optional[str] = None,
     title: Optional[str] = None,
     size: Literal[None, 4, 6] = None,
@@ -122,9 +129,13 @@ def circle_anchor(
     **kwargs: Any,
 ) -> SafeText:
     """Render a circular anchor link."""
-    if not href:
-        raise ValueError("Empty href supplied")
-    url = reverse(href, args=args, kwargs=kwargs)
+    match href:
+        case "":
+            raise ValueError("Empty href supplied")
+        case str():
+            url = reverse(href, args=args, kwargs=kwargs)
+        case model:
+            url = model.get_absolute_url()
     return format_html(
         '<a href="{url}" aria-label="{label}"{title} class="inline-block shrink-0 size-8 p-1.5 rounded-full border border-transparent hover:bg-secondary-hover active:bg-disabled">{icon}</a>',
         url=f"{url}#{fragment}" if fragment else url,
@@ -330,3 +341,9 @@ def picture(
         if fetchpriority
         else "",
     )
+
+
+@register.inclusion_tag("workspace/common/search_form.html")
+def workspace_search(workspace: Workspace) -> dict[str, object]:
+    """Render a workspace search form for the given workspace."""
+    return {"workspace": workspace, "form": WorkspaceSearchForm()}